% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gapclosing.R
\name{print.gapclosing}
\alias{print.gapclosing}
\title{Print function for gapclosing objects}
\usage{
\method{print}{gapclosing}(
  x,
  ...,
  digits = 2,
  quote = FALSE,
  right = FALSE,
  row.names = FALSE,
  max = NULL
)
}
\arguments{
\item{x}{An object of class \code{gapclosing}, which results from a call to the function \code{gapclosing}}

\item{...}{Other arguments to \code{print} commands}

\item{digits}{Argument passed to print.data.frame}

\item{quote}{Argument passed to print.data.frame}

\item{right}{Argument passed to print.data.frame}

\item{row.names}{Argument passed to print.data.frame}

\item{max}{Argument passed to print.data.frame}
}
\value{
Prints a summary of the estimates.
}
\description{
Prints the same output as generated by \code{summary}
}
\references{
Lundberg, Ian. 2021. "The gap-closing estimand: A causal approach to study interventions that close disparities across social categories." {https://osf.io/gx4y3/}
}
