% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gapclosing.R
\name{summary.gapclosing}
\alias{summary.gapclosing}
\title{Summary function for gapclosing objects}
\usage{
\method{summary}{gapclosing}(object, ...)
}
\arguments{
\item{object}{An object of class \code{gapclosing}, which results from a call to the function \code{gapclosing}}

\item{...}{Other arguments to \code{summary} commands}
}
\value{
Prints a summary of the estimates.
}
\description{
Summarizes the S3 class object returned by the gapclosing function
}
\references{
Lundberg, Ian. 2021. "The gap-closing estimand: A causal approach to study interventions that close disparities across social categories." {https://osf.io/gx4y3/}
}
