% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_estimator.R
\name{point_estimator}
\alias{point_estimator}
\title{Point estimator for gap-closing estimands}
\usage{
point_estimator(
  data_learn,
  data_estimate,
  counterfactual_assignments,
  outcome_formula,
  treatment_formula,
  category_name,
  outcome_name,
  treatment_name,
  treatment_algorithm = "glm",
  outcome_algorithm = "lm",
  weight_name = NULL
)
}
\arguments{
\item{data_learn}{Data frame in which treatment and outcome models will be learned}

\item{data_estimate}{Data frame in which the learned models will be converted to an estimate of the gap-closing estimand}

\item{counterfactual_assignments}{Numeric scalar or vector of length nrow(data), each element of which is on the [0,1] interval. If a scalar, the counterfactual probability by which all units are assigned to treatment condition 1. If a vector, each element i corresponds to the counterfactual probability by which each unit i is assigned to treatment condition 1.}

\item{outcome_formula}{Model formula the outcome. Covariates should include those needed for causal identification of the treatment effect (e.g. as defended in your Directed Acyclic Graph). If \code{outcome_algorithm} = "ranger", then the outcome model will be fit separately on the treatment and control groups. Otherwise, the user must specify all interactions in the formula.}

\item{treatment_formula}{Treatment formula, in the style formula(treatment ~ covariates). Covariates should include those needed for causal identification of the treatment effect (e.g. as defended in your Directed Acyclic Graph).}

\item{category_name}{Character name of the variable indicating the categories over which the gap is defined. Must be the name of a column in \code{data}.}

\item{outcome_name}{Character name of the outcome variable. Only required when there is no outcome_formula; otherwise extracted automatically. Must be a name of a column in \code{data}.}

\item{treatment_name}{Character name of the treatment variable. Only required when there is no treatment_formula; otherwise extracted automatically. Must be a name of a column in \code{data}.}

\item{treatment_algorithm}{Character name of the algorithm for the treatment model. One of "glm", "ridge", "gam", or "ranger". Defaults to "glm", which is a logit model. Option "ridge" is ridge regression. Option "gam" is a generalized additive model fit (see package \code{mgcv}). Option "ranger" is a random forest (see package \code{ranger}). If "ranger", this function avoids propensity scores equal to 0 or 1 by bottom- and top-coding predicted values at .001 and .999.}

\item{outcome_algorithm}{Character name of the algorithm for the outcome model. One of "lm", "ridge", "gam", or "ranger". Defaults to "lm", which is an OLS model. Option "ridge" is ridge regression. Option "gam" is a generalized additive model fit (see package \code{mgcv}). Option "ranger" is a random forest (see package \code{ranger}).}

\item{weight_name}{Character name of a sampling weight variable, if any, which captures the inverse probability of inclusion in the sample. The default assumes a simple random sample (all weights equal).}
}
\value{
@return A list with four elements.
\itemize{
\code{counterfactual_means} A tibble with a counterfactual mean estimate for each category
\code{counterfactual_means} A tibble with a counterfactual disparity estimate for each pair of categories
\code{treatment_model} Object containing the fitted treatment model
\code{outcome_model} Object containing the fitted outcome model
}
}
\description{
This is an internal function typically called from other functions rather than by the user. It uses a learning sample to learn the nuisance functions (treatment and outcome model) and then an auxiliary estimation sample to use those functions in estimation of the gap-closing estimand. For single-sample estimation, both the learning and estimation samples are the same. For cross-fitting, this function is called repeatedly with the roles of each dataset swapped.
}
\references{
Lundberg I (2021). "The gap-closing estimand: A causal approach to study interventions that close disparities across social categories." Sociological Methods and Research. Available at {https://osf.io/gx4y3/}.
}
